// lib/types.ts
export interface Order {
  _id: string;
  customerId: string;
  orderAmount: number;
  orderDate: string;
  createdAt: string;
  updatedAt: string;
}